// apps/web/core/components/templates/settings/project/form/modules/types.ts
/**
 * SPDX-FileCopyrightText: 2023-present Plane Software, Inc.
 * SPDX-License-Identifier: LicenseRef-Plane-Commercial
 *
 * Licensed under the Plane Commercial License (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * https://plane.so/legals/eula
 *
 * DO NOT remove or modify this notice.
 * NOTICE: Proprietary and confidential. Unauthorized use or distribution is prohibited.
 */

import type { TModuleStatus } from "@plane/types";

export type TModuleFormValues = {
  name: string;
  description: string;
  status: TModuleStatus;
  lead_id: string | null;
  member_ids: string[];
};